LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/semantics/disabled-elements/disabled-event-dispatch.tentative-expected.txt

hello world child  hello world child

PASS Testing mousedown events when clicking child of disabled button.
PASS Testing mousedown events when clicking child of disabled my-control.
PASS Testing mousedown events when clicking disabled button.
PASS Testing mousedown events when clicking disabled my-control.
PASS Testing mouseup events when clicking child of disabled button.
PASS Testing mouseup events when clicking child of disabled my-control.
PASS Testing mouseup events when clicking disabled button.
PASS Testing mouseup events when clicking disabled my-control.
FAIL Testing pointerdown events when clicking child of disabled button. assert_equals: parent element received pointerdown events expected true but got false
FAIL Testing pointerdown events when clicking child of disabled my-control. assert_equals: parent element received pointerdown events expected true but got false
FAIL Testing pointerdown events when clicking disabled button. assert_equals: parent element received pointerdown events expected true but got false
FAIL Testing pointerdown events when clicking disabled my-control. assert_equals: parent element received pointerdown events expected true but got false
FAIL Testing pointerup events when clicking child of disabled button. assert_equals: parent element received pointerup events expected true but got false
FAIL Testing pointerup events when clicking child of disabled my-control. assert_equals: parent element received pointerup events expected true but got false
FAIL Testing pointerup events when clicking disabled button. assert_equals: parent element received pointerup events expected true but got false
FAIL Testing pointerup events when clicking disabled my-control. assert_equals: parent element received pointerup events expected true but got false
PASS Testing click events when clicking child of disabled button.
PASS Testing click events when clicking child of disabled my-control.
PASS Testing click events when clicking disabled button.
PASS Testing click events when clicking disabled my-control.
PASS Testing dblclick events when clicking child of disabled button.
PASS Testing dblclick events when clicking child of disabled my-control.
PASS Testing dblclick events when clicking disabled button.
PASS Testing dblclick events when clicking disabled my-control.
